requirements.txt
flask>=3.0.0
openai>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
transformers>=4.38.0
torch>=2.1.0
faiss-cpu>=1.7.4
ultralytics>=8.2.0
sentencepiece>=0.1.99
protobuf>=4.25.0
